+++ src/main/java/com/seveneleven/devlens/domain/project/service/ProjectServiceImpl.java
@@ -1,14 +1,11 @@
 package com.seveneleven.devlens.domain.project.service;
 
-import com.seveneleven.devlens.domain.member.constant.MemberStatus;
-import com.seveneleven.devlens.domain.member.repository.MemberRepository;
 import com.seveneleven.devlens.domain.member.service.MemberService;
 import com.seveneleven.devlens.domain.project.dto.GetProjectDetail;
 import com.seveneleven.devlens.domain.project.dto.GetProjectList;
+import com.seveneleven.devlens.domain.project.service.dashboard.ProjectReader;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
-
-import java.util.List;
 
 @Service
 @RequiredArgsConstructor
@@ -20,7 +17,6 @@ public class ProjectServiceImpl implements ProjectService {
     public GetProjectList.Response getProjectList(Long memberId) {
         // TODO - securityUtils 사용 확인
         // TODO - MemberRepository -> Reader로 변경 필요
-
         return projectReader.getProjectList(memberId, memberService.getCompanyIdById(memberId));
     }
